backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
dj-rest-auth>=5.0
django-allauth>=0.61
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
razorpay>=1.4
reportlab>=4.1
python-dotenv>=1.0
django-filter>=23.5
Pillow>=10.2
